nzxt/requirements.txt
psutil>=5.9.0
requests>=2.28.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
opencv-python>=4.6.0
pygame>=2.1.0
colorama>=0.4.5
pyaudio>=0.2.11
librosa>=0.9.2
scipy>=1.7.0
sounddevice>=0.4.5
webrtcvad>=2.0.10
yt-dlp>=2023.3.4
